refactor: 컨벤션에 맞도록 수정
- prettier 적용
- 파일 네이밍 소문자로 통일
